move edittab form to usetab, changed theme reducer to no longer be hard coded
